# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ICreaBomOutput": {
				"properties": {
					"error": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreaBomInput": {
				"properties": {
					"id_obra": {
						"type": "number",
						"format": "double"
					},
					"reserva": {
						"type": "number",
						"format": "double"
					},
					"materiales": {
						"type": "string"
					}
				},
				"required": [
					"id_obra",
					"reserva",
					"materiales"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConsultaBomOutput": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"id_obra": {
						"type": "number",
						"format": "double"
					},
					"cod_reserva": {
						"type": "number",
						"format": "double"
					},
					"codigo_sap_material": {
						"type": "number",
						"format": "double"
					},
					"cantidad_requerida": {
						"type": "number",
						"format": "double"
					},
					"fecha_ingreso": {
						"type": "string"
					},
					"rut_usuario": {
						"type": "string"
					},
					"persona": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"id_obra",
					"cod_reserva",
					"codigo_sap_material",
					"cantidad_requerida",
					"fecha_ingreso",
					"rut_usuario",
					"persona"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMensajeHome": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"mensaje": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"mensaje"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IHomePage": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"routerlink": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"routerlink"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IMenuItem.Exclude_keyofIMenuItem.orden__": {
				"properties": {
					"label": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"routerLink": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"icon",
					"routerLink"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IMenuItem.orden_": {
				"$ref": "#/components/schemas/Pick_IMenuItem.Exclude_keyofIMenuItem.orden__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IJsonMenu": {
				"properties": {
					"label": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/Omit_IMenuItem.orden_"
						},
						"type": "array"
					}
				},
				"required": [
					"label",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRespuestaLogin": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"nombre": {
						"type": "string"
					},
					"funcion": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mensaje": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IMensajeHome"
							}
						],
						"nullable": true
					},
					"homepage": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IHomePage"
							}
						],
						"nullable": true
					},
					"accessToken": {
						"type": "string"
					},
					"menu": {
						"items": {
							"$ref": "#/components/schemas/IJsonMenu"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"username",
					"nombre",
					"funcion",
					"email",
					"roles",
					"accessToken",
					"menu"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignInInput": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRespuestaCambioPass": {
				"properties": {
					"error": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICambioPassInput": {
				"properties": {
					"password": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUserRepository.Exclude_keyofIUserRepository.password__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"sistema": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUserRepository.password_": {
				"$ref": "#/components/schemas/Pick_IUserRepository.Exclude_keyofIUserRepository.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "auth-token",
				"in": "header",
				"description": "The authorization token"
			}
		}
	},
	"info": {
		"title": "backend-pelom",
		"version": "1.0.0",
		"description": "Backend para Webapp Pelom",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Héctor Miranda"
		}
	},
	"paths": {
...
		"/obras/backoffice/v1/creabom": {
			"post": {
				"operationId": "CreateBomMasivo",
				"responses": {
					"200": {
						"description": "ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreaBomOutput"
								}
							}
						}
					}
				},
				"description": "Función de creación de BOM",
				"tags": [
					"Obras - Backoffice - Manejo materiales (bom)"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreaBomInput"
							}
						}
					}
				}
			}
		},
		"/obras/backoffice/v1/bom_inicial_por_obra": {
			"get": {
				"operationId": "GetBomZero",
				"responses": {
					"200": {
						"description": "ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IConsultaBomOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Obras - Backoffice - Manejo materiales (bom)"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id_obra",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/obras/backoffice/v1/bom_actual_por_obra": {
			"get": {
				"operationId": "GetBomFinal",
				"responses": {
					"200": {
						"description": "ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IConsultaBomOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Obras - Backoffice - Manejo materiales (bom)"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id_obra",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/v1/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"201": {
						"description": "usuario ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRespuestaLogin"
								}
							}
						}
					}
				},
				"description": "Función de Login",
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignInInput"
							}
						}
					}
				}
			}
		},
...
		"/auth/v1/cambiapassword": {
			"post": {
				"operationId": "CambioPassword",
				"responses": {
					"200": {
						"description": "consulta ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRespuestaCambioPass"
								}
							}
						}
					}
				},
				"tags": [
					"Autenticación"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICambioPassInput"
							}
						}
					}
				}
			}
		},
...
		"/auth/v1/me": {
			"post": {
				"operationId": "Me",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Omit_IUserRepository.password_"
								}
							}
						}
					}
				},
				"description": "Get the current user.",
				"tags": [
					"Autenticación"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:7080/api"
		}
	]
}